+++ lib/clixon/clixon_netconf_lib.h
@@ -63,5 +63,6 @@ int netconf_operation_failed_xml(cxobj **xret, char *type, char *message);
 int netconf_malformed_message(cbuf *cb, char *message);
 int netconf_malformed_message_xml(cxobj **xret, char *message);
 int netconf_trymerge(cxobj *x, yang_spec *yspec, cxobj **xret);
+int netconf_module_load(clicon_handle h);
 
 #endif /* _CLIXON_NETCONF_LIB_H */